Don't check for a private function.
Even with WordPress’ commitment to backwards compatibility, we should
probably not bank on private functions being around for ever, and tying
our functionality to it.

This will go old school and do a version check to provide compatibility
with older versions of WordPress. Again, this is only temporary until
4.3 is out.

See #644.
Fixes #636.